refactor!: remove LedgerModal, IssuanceModal and dependent code
