+++ web-components/src/components/modal/modal.stories.tsx
@@ -8,7 +8,6 @@ import { BasketTakeModal } from './BasketTakeModal';
 import { ConfirmModal } from './ConfirmModal';
 import { CreateSellOrderModal } from './CreateSellOrderModal';
 import CropImageModal from './CropImageModal';
-import IssuanceModal from './IssuanceModal';
 import { ProcessingModal } from './ProcessingModal';
 import { TxErrorModal } from './TxErrorModal';
 import { TxSuccessfulModal } from './TxSuccessfulModal';
@@ -28,76 +27,6 @@ function OpenModal(): JSX.Element {
       <Modal open={open} onClose={() => setOpen(false)}>
         <p>I'm a Modal</p>
       </Modal>
-    </div>
-  );
-}
-
-function OpenLedgerModal(): JSX.Element {
-  const [open, setOpen] = useState(false);
-  return (
-    <div>
-      <Button variant="contained" onClick={() => setOpen(true)}>
-        Open Modal
-      </Button>
-      <IssuanceModal
-        open={open}
-        onClose={() => setOpen(false)}
-        link="https://devnet.regen.aneka.io/txs/C7F4DDC696BB64605D2ACB45D5999CF3E36F2F73783E6DB5066CFB282A6E7C42"
-        issuer={{
-          name: 'Regen Network Development, Inc.',
-          type: 'ORGANIZATION',
-          address: 'regen:1p6syuqk3e5a8hwp8e20jyjwr8p7nj270x4spqm',
-          location: 'Great Barrington, Massachussetts, United States',
-          description:
-            'Regen Network realigns the agricultural economy with ecological health by creating the global marketplace for planetary stewardship.',
-        }}
-        issuees={[
-          {
-            name: 'Wyelba Pty Ltd',
-            type: 'ORGANIZATION',
-            address: 'regen:1wuufq6vkl4qmmgzs06mtgatklpxr5zr4qqnk4k',
-            location: 'New South Wales, Australia',
-            description: 'Lorem ipsum',
-          },
-        ]}
-        timestamp={new Date('2020-12-16 09:23:41.904137+00')}
-        numberOfCredits={28682}
-        creditUnit="1 ton of CO2e"
-        vintageId={{
-          name: 'certificate',
-          info: '634aabf4',
-          link: 'https://regen-registry.s3.amazonaws.com/projects/wilmot/wilmot-certificate-121620.pdf',
-        }}
-        vintagePeriod="2017-2019"
-        monitoringPeriods={[
-          {
-            name: 'monitoring report',
-            info: '2017-2018',
-            link: 'https://regen-registry.s3.amazonaws.com/projects/wilmot/Wilmot+Monitoring+Report+2018.pdf',
-          },
-          {
-            name: 'monitoring report',
-            info: '2018-2019',
-            link: 'https://regen-registry.s3.amazonaws.com/projects/wilmot/Wilmot+Monitoring+Report+2019.pdf',
-          },
-        ]}
-        projectName="Wilmot"
-        standard={{
-          documentId: 'RND_PG',
-          name: 'Regen Registry Program Guide',
-          version: 'v1.0',
-        }}
-        creditClass={{
-          documentId: 'RND_CC_0001',
-          name: 'CarbonPlus Grasslands',
-          version: 'v0.9',
-        }}
-        methodology={{
-          documentId: 'RND_M0001',
-          name: 'CarbonPlus Grasslands',
-          version: 'v0.9',
-        }}
-      />
     </div>
   );
 }
@@ -187,7 +116,6 @@ function OpenCropImageModal(props: CropStoryProps): JSX.Element {
 }
 
 export const modal = (): JSX.Element => <OpenModal />;
-export const ledgerModal = (): JSX.Element => <OpenLedgerModal />;
 export const cropSquareImageModal = (): JSX.Element => <OpenCropImageModal />;
 export const cropRoundImageModal = (): JSX.Element => (
   <OpenCropImageModal circularCrop />
